Remove old updateCardTemplateTagNames, move and refactor relevant test to initializeQB unit test suite
